requirements.txt
black
gitpython
isort
jupyter
laion-clap
librosa
matplotlib
numpy<2.0
pandas
pip
scaper
scikit-learn
scipy
sox
soundata
soundfile
torch
torchaudio
torchlibrosa
torchvision
tqdm
wget